% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomaRt_annotation_mus_musculus}
\alias{biomaRt_annotation_mus_musculus}
\title{biomaRt_annotation_mus_musculus}
\format{
An object of class data frame
}
\source{
data-raw/biomaRt_annotation_tables_based_on_GTFs_ensembl110.R
}
\usage{
biomaRt_annotation_mus_musculus
}
\description{
Biomart annotation table for mus musculus
}
\keyword{datasets}
