% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{export_data}
\alias{export_data}
\title{Export Data Table}
\usage{
export_data(comparison_name)
}
\arguments{
\item{comparison_name}{A DESeq2 results object.}
}
\value{
A data frame ready for export. It includes information about the compared samples
and the parameters used for DESeq2 analysis.
}
\description{
This function prepares the results of the differential expression analysis for export,
adding the sample names and the parameters used in the DESeq2 analysis.
}
