% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostic_plots.R
\name{plot_diagnostic_plots}
\alias{plot_diagnostic_plots}
\title{Diagnostic Plots for Count Matrix and Metadata}
\usage{
plot_diagnostic_plots(count_matrix_raw, metadata_raw)
}
\arguments{
\item{count_matrix_raw}{A matrix where rows represent genes and columns represent samples.}

\item{metadata_raw}{A dataframe containing metadata information for the samples.}
}
\value{
This function saves the generated plots as .html or .png files in your working directory.
}
\description{
This function generates several diagnostic plots based on input count matrix and metadata.
It includes PCA plot, Boxplot, and Density plot. The function also normalizes counts using DESeq2's median of ratios method.
}
